+++ crates/next-core/js/src/entry/config/next.ts
@@ -3,18 +3,16 @@ import loadConfig from "next/dist/server/config";
 import { PHASE_DEVELOPMENT_SERVER } from "next/dist/shared/lib/constants";
 
 type IpcIncomingMessage = {
-  type: "loadNextConfig" | "loadPostcssConfig";
+  type: "loadNextConfig";
   path: string;
 };
 
 type IpcOutgoingMessage = {
   type: "javaScriptValue";
-  data: number[];
+  data: string;
 };
 
 const ipc = IPC as Ipc<IpcIncomingMessage, IpcOutgoingMessage>;
-
-const CACHED_CONFIGS = new Map<string, any>();
 
 (async () => {
   while (true) {
@@ -23,14 +21,13 @@ const CACHED_CONFIGS = new Map<string, any>();
     switch (msg.type) {
       case "loadNextConfig": {
         const nextConfig = await loadConfig(PHASE_DEVELOPMENT_SERVER, msg.path);
-        CACHED_CONFIGS.set(msg.path, nextConfig);
         // @ts-expect-error
         nextConfig.rewrites = await nextConfig.rewrites?.();
         // @ts-expect-error
         nextConfig.redirects = await nextConfig.redirects?.();
         await ipc.send({
           type: "javaScriptValue",
-          data: Array.from(Buffer.from(JSON.stringify(nextConfig))),
+          data: JSON.stringify(nextConfig),
         });
         break;
       }
